:truck: move MGazine to another file
